Back button support, waiting class, performance improvements.
Added instanceId and $.fn.ajaxify.instanceCount variables so that we can track the instances of ajaxify that have been called. Use this rather than stringifying settings to see which instance should handle a given load. Add back button support - we'll do a generic, broad animation to make sure we get everything. If the user hits a page that hasn't been ajax loaded (i.e., she visited the URL directly), the first instance of ajaxify will handle that load. In the default animations, there's an 'ajaxify-waiting' class that appears between the time that the 'begin' animation completes and the time that the 'end' animation is triggered. With this class, you can display a loading spinner only if the animation takes a while, and there's a delay between the two animations.