Ajaxify now runs on the HTML element, rather than an element selected by jQuery.
This simplifies how the plugin is called. It also means that you can use body classes in your selectors. Selecting an element and calling ajaxify on it introduced the potential for errors - for example, you could call it on 'body.my-page', and if the user navigated from body.another-page to body.my-page, ajaxify would not work for any links that were clicked on body.my-page.